// frontend/src/types/supplier.ts
export type Supplier = {
  id: number;

  supplier_name: string;

  company_name: string;

  contact_person: string | null;

  phone: string;

  email: string | null;

  address: string | null;

  city: string | null;

  state: string | null;

  country: string | null;

  gst_number: string | null;
};